FEAT add stv counting algorithm
stv algorithm goes after shuffle
