+++ app/psifos/crypto/tally/tally.py
@@ -5,15 +5,16 @@ Tally module for Psifos.
 from app.database.serialization import SerializableList
 from app.psifos.psifos_object.result import ElectionResult
 from .homomorphic.tally import HomomorphicTally
-from .mixnet.tally import MixnetTally
+from .mixnet.close_massive_tally import CloseMassiveTally
+from .mixnet.stv_tally import STVTally
 
 class TallyFactory():
     @staticmethod
     def create(**kwargs):
         tally_to_mn_tally = {
             "homomorphic":HomomorphicTally,
-            "mixnet":MixnetTally,
-            "stvnc":MixnetTally,
+            "mixnet":CloseMassiveTally,
+            "stvnc":STVTally,
         }
         tally_type = kwargs.get("tally_type")
         if tally_type in tally_to_mn_tally.keys():
