+++ requirements.txt
@@ -65,6 +65,7 @@ sniffio==1.3.0
 SQLAlchemy==1.4.40
 starlette==0.27.0
 starlette-context==0.3.4
+STVPoll==0.4.5
 toml==0.10.2
 tomli==2.0.1
 tornado==6.3.3
